# apps/web/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zero Mail API",
    "description": "Phase 1 skeleton + Phase 1.1 i18n/error contract",
    "version": "0.1.1"
  },
  "servers": [
    {
      "url": "http://localhost:59080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/llm/byok": {
      "get": {
        "tags": ["llm-byok"],
        "operationId": "current",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ByokCurrentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["llm-byok"],
        "operationId": "save",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ByokSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ByokSaveResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/llm/byok/validate": {
      "post": {
        "tags": ["llm-byok"],
        "operationId": "validate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ByokValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ByokValidateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/me/language": {
      "patch": {
        "tags": ["me-controller"],
        "operationId": "updateLanguage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLanguageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/me": {
      "get": {
        "tags": ["me-controller"],
        "operationId": "me",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ByokSaveRequest": {
        "type": "object",
        "properties": {
          "preset": {
            "type": "string",
            "enum": [
              "openrouter",
              "openai",
              "anthropic",
              "google-genai",
              "deepseek",
              "openai-compatible",
              "anthropic-compatible"
            ]
          },
          "endpoint": {
            "type": "string"
          },
          "model": {
            "type": "string",
            "minLength": 1
          },
          "apiKey": {
            "type": "string"
          }
        },
        "required": ["apiKey", "model", "preset"]
      },
      "ByokSaveResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "savedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ByokValidateRequest": {
        "type": "object",
        "properties": {
          "preset": {
            "type": "string",
            "enum": [
              "openrouter",
              "openai",
              "anthropic",
              "google-genai",
              "deepseek",
              "openai-compatible",
              "anthropic-compatible"
            ]
          },
          "endpoint": {
            "type": "string"
          },
          "model": {
            "type": "string",
            "minLength": 1
          },
          "apiKey": {
            "type": "string"
          }
        },
        "required": ["apiKey", "model", "preset"]
      },
      "ByokValidateResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "models": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reason": {
            "type": "string"
          }
        }
      },
...
      "UpdateLanguageRequest": {
        "type": "object",
        "properties": {
          "language": {
            "type": "string",
            "minLength": 1,
            "pattern": "vi|en"
          }
        },
        "required": ["language"]
      },
      "GmailConnectionStatusExtended": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "ingestionHealth": {
            "type": "string"
          },
          "googleEmail": {
            "type": "string"
          }
        }
      },
      "MeResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "onboardingStep": {
            "type": "string"
          },
          "preferredLanguage": {
            "type": "string"
          },
          "triagePaused": {
            "type": "boolean"
          },
          "gmailConnectionStatus": {
            "$ref": "#/components/schemas/GmailConnectionStatusExtended"
          }
        }
      },
...
      "ByokCurrentResponse": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "enum": ["anthropic", "deepseek", "google-genai", "openai"]
          },
          "endpointHost": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "savedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ApiError": {
        "type": "object",
        "description": "RFC 9457 ProblemDetail extended with code/params/fieldErrors. Wire shape produced by Spring 7 ProblemDetail.setProperty(...) flattened by Jackson 3 ProblemDetailJacksonMixin. Phase 1.1 introduces this contract; the FE switches on `code`.",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "description": "RFC 9457 problem type URI; defaults to `about:blank`."
          },
          "title": {
            "type": "string",
            "description": "Generic English diagnostic. NOT the user-facing message — FE ignores it."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code echoed in the body."
          },
          "detail": {
            "type": "string",
            "description": "Generic English diagnostic. NOT the user-facing message — FE ignores it."
          },
          "instance": {
            "type": "string",
            "format": "uri",
            "description": "Optional URI identifying this specific occurrence."
          },
          "code": {
            "type": "string",
            "description": "Stable dotted hierarchical key (e.g. `error.auth.unauthorized`). Frontend localizes via the errors.* dictionary. ALWAYS PRESENT."
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "description": "Allow-listed scalar value: string, integer, number, or boolean. Filtered through AllowedParamScalars before serialization.",
              "oneOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer",
                  "format": "int32"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "description": "Allow-listed ICU placeholder scalars for the FE error message. Filtered through AllowedParamScalars; never carries raw user content, Gmail body fragments, SQL constraint names, or exception class names."
          },
          "fieldErrors": {
            "type": "array",
            "description": "Per-field validation errors. Present only on validation failures.",
            "items": {
              "$ref": "#/components/schemas/FieldErrorDto"
            }
          },
          "message": {
            "type": "string",
            "deprecated": true,
            "description": "Transitional alias of `code`, kept for one phase per CONTEXT.md decision D-C1. Prefer `code`."
          }
        },
        "required": ["code"]
      },
      "FieldErrorDto": {
        "type": "object",
        "description": "Per-field validation error. CONTEXT.md decision D-C2: { field, code, params }. The JHipster `message` and `objectName` fields are deliberately rejected — the frontend localizes the dotted `code` via the errors.field.* dictionary.",
        "properties": {
          "field": {
            "type": "string",
            "description": "Dotted path to the offending field, e.g. `email`, `templates[0].key`."
          },
          "code": {
            "type": "string",
            "description": "Dotted hierarchical key inside the errors.field.* namespace."
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "description": "Allow-listed scalar value: string, integer, number, or boolean. Filtered through AllowedParamScalars before serialization.",
              "oneOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer",
                  "format": "int32"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "description": "Allow-listed scalars (numbers, booleans, dotted keys, short resource identifiers). Never raw user content."
          }
        },
        "required": ["code", "field"]
      }
    }
  }
}