# backend/api/src/main/java/com/zeromail/api/dto/llm/ByokValidateResponse.java
package com.zeromail.api.dto.llm;

import java.util.List;

import com.zeromail.core.llm.application.ByokValidateResult;

public record ByokValidateResponse(boolean ok, List<String> models, String reason) {

  public ByokValidateResponse {
    models = models == null ? null : List.copyOf(models);
  }

  public static ByokValidateResponse from(ByokValidateResult result) {
    return new ByokValidateResponse(result.ok(), result.models(), result.reason());
  }
}